+++ NEWS.md
@@ -1,3 +1,9 @@
+# qrcode 0.2.1
+
+* Replace the unexported `qr_logo()` by `add_logo`().
+* `add_logo()` returns a `qr_code` object.
+* `plot()` and `generate_svg()` handle the output of `add_logo()`.
+
 # qrcode 0.2.0
 
 * Defunct legacy functions.